* develope: (65 commits)
  Add sonarqube permission Fix put project parameters
  API add passing_rate_detail update sync_redmine.py response sorted by project_id
  get_project_member_detail remove limit(10)
  add project_members_count, project_members_detail update sync_redmine multiple insert style
  update remine project table column
  update apis/resources/sync_redmine.py
  api.py update /dashboard/project_members url path
  DB git_commit_number_each_days add "created_at" column
  table GitCommitNumberEachDays add created_at column
  add standard api output
  Add /sync_gitlab/count_each_pj_commits_by_days api to count each project commit time.
  table remove foreign key repo_id and add repo_name
  add first version.
  dev
  bug fix. add "skip_reconfirmation: true" in gitlab put /user/{user_id} API
  fix PUT /user/{user_id}, need to email verify bug.
  Create GItCommitNumberEashDays table
  when API [GET] /dashboard/the_last_hours_commits get commit info, ignore "IIIdevops-templates" group
  add sync_redmine class API
  RedmineIssue add assigned_to_id ProjectMember add id, user_id, project_id IssueRank adjust project_count
  ...

+++ apis/resources/project.py
@@ -4,9 +4,9 @@ import base64
 
 from flask_jwt_extended import jwt_required, get_jwt_identity
 from flask_restful import Resource, reqparse
+from kubernetes.client import ApiException
 from sqlalchemy import desc
 from sqlalchemy.orm.exc import NoResultFound
-
 import config
 import model
 import nexus
@@ -39,20 +39,9 @@ def list_projects(user_id):
     project_id_list = []
     for row in rows:
         project_id_list.append(row.Project.id)
-
-    pm_map = {}
-    pms = db.session.query(model.User, model.Project.id) \
-        .join(model.ProjectUserRole,
-              model.ProjectUserRole.user_id == model.User.id) \
-        .filter(model.ProjectUserRole.project_id.in_(project_id_list),
-                model.ProjectUserRole.role_id.in_((role.PM.id, role.ADMIN.id))) \
-        .all()
-    for pm in pms:
-        pm_map[pm.id] = pm.User
-
+        
     projects = redmine.rm_list_projects()
     issues = redmine.rm_list_issues()
-
     output_array = []
     for row in rows:
         project_id = row.Project.id
@@ -73,21 +62,21 @@ def list_projects(user_id):
             if issue["status"]["name"] == "Closed":
                 closed_count += 1
             if issue["due_date"] is not None:
-                if (datetime.today() > datetime.strptime(
+                if (datetime.utcnow() > datetime.strptime(
                         issue["due_date"], "%Y-%m-%d")):
                     overdue_count += 1
             total_count += 1
             del issue
-
         project_status = "進行中"
         if total_count == 0:
             project_status = "未開始"
         if closed_count == total_count and total_count != 0:
             project_status = "已結案"
 
-        pm = pm_map[project_id]
-        if pm is None:
+        if row.Project.owner_id is None:
             pm = model.User(id=0, name='No One')
+        else:
+            pm = nexus.nx_get_user(row.Project.owner_id)
 
         updated_on = None
         for pjt in projects:
@@ -114,7 +103,9 @@ def list_projects(user_id):
             "project_status": project_status,
             "closed_count": closed_count,
             "total_count": total_count,
-            "overdue_count": overdue_count
+            "overdue_count": overdue_count,
+            'start_date' : str(row.Project.start_date),
+            'due_date' : str(row.Project.due_date)
         })
 
     return util.success({"project_list": output_array})
@@ -128,14 +119,16 @@ def create_project(user_id, args):
     if args['display'] is None:
         args['display'] = args['name']
     project_name = args['name']
-
     # create namespace in kubernetes
     try:
         kubernetesClient.create_namespace(project_name)
         kubernetesClient.create_role_in_namespace(project_name)
         kubernetesClient.create_namespace_quota(project_name)
         kubernetesClient.create_namespace_limitrange(project_name)
-    except Exception as e:
+    except ApiException as e:
+        if e.status == 409:
+            raise DevOpsError(e.status, 'Kubernetes already has this identifier.',
+                              error=apiError.identifier_has_been_taken(args['name']))
         kubernetesClient.delete_namespace(project_name)
         raise e
 
@@ -204,7 +197,7 @@ def create_project(user_id, args):
                         if len(resp['errors']) > 0:
                             if resp['errors'][0] == 'Identifier has already been taken':
                                 raise DevOpsError(status_code, 'Redmine already used this identifier.',
-                                                  error=apiError.identifier_has_been_token(args['name']))
+                                                  error=apiError.identifier_has_been_taken(args['name']))
                     raise e
                 elif service == 'gitlab':
                     status_code = e.status_code
@@ -214,7 +207,7 @@ def create_project(user_id, args):
                             if gitlab_json['message']['name'][0] == 'has already been taken':
                                 raise DevOpsError(
                                     status_code, {"gitlab": gitlab_json},
-                                    error=apiError.identifier_has_been_token(args['name'])
+                                    error=apiError.identifier_has_been_taken(args['name'])
                                 )
                         except (KeyError, IndexError):
                             pass
@@ -239,7 +232,12 @@ def create_project(user_id, args):
             description=args['description'],
             ssh_url=gitlab_pj_ssh_url,
             http_url=gitlab_pj_http_url,
-            disabled=args['disabled']
+            disabled=args['disabled'],
+            start_date = args['start_date'],
+            due_date = args['due_date'],
+            create_at = str(datetime.utcnow()),
+            owner_id = user_id,
+
         )
         db.session.add(new_pjt)
         db.session.commit()
@@ -315,48 +313,29 @@ def create_bot(project_id):
         project_id, 'nexus-bot', {'username': login, 'password': password})
 
 
+def check_modify_database_type(items, args, select_db):    
+    output = {}
+    for item in items:
+        if args[item] is None:
+             output[item] = getattr(select_db, item)
+        else:
+            output[item] = args[item]
+            setattr(select_db,item,args[item])
+    return args , select_db
+
 @record_activity(ActionType.UPDATE_PROJECT)
 def pm_update_project(project_id, args):
-    result = db.engine.execute(
-        "SELECT * FROM public.project_plugin_relation WHERE project_id = '{0}'".format(
-            project_id))
-    project_relation = result.fetchone()
-    result.close()
-
-    redmine_project_id = project_relation["plan_project_id"]
-    gitlab_project_id = project_relation["git_repository_id"]
-
-    if args["name"] is None:
-        result = db.engine.execute(
-            "SELECT name FROM public.projects WHERE id = '{0}'".format(
-                project_id))
-        args["name"] = result.fetchone()[0]
-        result.close()
-    if args["display"] is None:
-        result = db.engine.execute(
-            "SELECT name FROM public.projects WHERE id = '{0}'".format(
-                project_id))
-        args["display"] = result.fetchone()[0]
-        result.close()
-    if args["description"] is None:
-        result = db.engine.execute(
-            "SELECT description FROM public.projects WHERE id = '{0}'".format(
-                project_id))
-        args["description"] = result.fetchone()[0]
-        result.close()
-
-    gitlab.gl_update_project(gitlab_project_id, args["description"])
-    redmine.rm_update_project(redmine_project_id, args)
-    # 修改db
-    # 修改projects
-    fields = ['name', 'display', 'description', 'disabled']
-    for field in fields:
-        if args[field] is not None:
-            db.engine.execute(
-                "UPDATE public.projects SET {0} = '{1}' WHERE id = '{2}'".format(
-                    field, args[field], project_id))
-
+    targets = ['display','name','description','disabled','owner_id','start_date', 'due_date']
+    plugin_relation = model.ProjectPluginRelation.query.filter_by(project_id=project_id).first()
+    if args['description'] is not None:
+        gitlab.gl_update_project(plugin_relation.git_repository_id, args["description"])
+    redmine.rm_update_project(plugin_relation.plan_project_id, args)
+    project = model.Project.query.filter_by(id=project_id).first()                
+    args, project = check_modify_database_type(targets,args, project)    
+    project.update_at = str(datetime.utcnow())
+    db.session.commit()
     return util.success()
+
 
 
 def try_to_delete(delete_method, argument):
@@ -664,7 +643,7 @@ def get_projects_by_user(user_id):
         if len(issue_due_date_list) != 0:
             next_d_time = min(
                 issue_due_date_list,
-                key=lambda d: abs(d - datetime.now()))
+                key=lambda d: abs(d - datetime.utcnow()))
         if next_d_time is not None:
             output_dict['next_d_time'] = next_d_time.isoformat()
 
@@ -970,10 +949,12 @@ class SingleProject(Resource):
         role.require_pm("Error while updating project info.")
         role.require_in_project(project_id, "Error while updating project info.")
         parser = reqparse.RequestParser()
-        parser.add_argument('name', type=str)
         parser.add_argument('display', type=str)
         parser.add_argument('description', type=str)
         parser.add_argument('disabled', type=bool)
+        parser.add_argument('start_date', type=str)
+        parser.add_argument('due_date', type=str)
+        parser.add_argument('owner_id', type=int)
         args = parser.parse_args()
         return pm_update_project(project_id, args)
 
@@ -995,8 +976,9 @@ class SingleProject(Resource):
         parser.add_argument('template_id', type=int)
         parser.add_argument('tag_name', type=str)
         parser.add_argument('arguments', type=dict)
+        parser.add_argument('start_date', type=str, required=True)
+        parser.add_argument('due_date', type=str, required=True)
         args = parser.parse_args()
-
         pattern = "^[a-z][a-z0-9-]{0,28}[a-z0-9]$"
         result = re.fullmatch(pattern, args["name"])
         if result is None:
@@ -1130,7 +1112,6 @@ class ProjectUserResourcePodLog(Resource):
         parser = reqparse.RequestParser()
         parser.add_argument('container_name', type=str)
         args = parser.parse_args()
-        print(project_id)
         return get_kubernetes_namespace_pod_log(project_id, pod_name, args['container_name'])
 
 class ProjectEnvironment(Resource):
